+++ packages/gateway/CHANGELOG.md
@@ -1,5 +1,15 @@
 # lyra-gateway
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - lyra-plugin-onchain@0.2.1
+  - lyra-core@0.2.1
+  - lyra-plugin-system@0.2.1
+  - lyra-plugin-telegram@0.2.1
+
 ## 0.2.0
 
 ### Patch Changes