+++ packages/plugin-telegram/CHANGELOG.md
@@ -1,5 +1,11 @@
 # lyra-plugin-telegram
 
+## 0.2.1
+
+### Patch Changes
+
+- lyra-core@0.2.1
+
 ## 0.2.0
 
 ### Patch Changes